refactor: moving the rewrite to the root of the repo
