--- crates/ratings_new/Cargo.toml
@@ -1,44 +1,0 @@
-[package]
-name = "ratings_new"
-version = "0.1.0"
-edition = "2021"
-
-[features]
-# Used to prevent tests that need a running DB from executing unless explicitly requested
-# Annotate such tests with the following:
-#   #[cfg_attr(not(feature = "db_tests"), ignore)]
-#
-# Execute the tests using "cargo test --features db_tests"
-db_tests = []
-
-[dependencies]
-dotenvy = "0.15"
-envy = "0.4"
-http = "1.1.0"
-jsonwebtoken = "9.2"
-prost = "0.13.3"
-prost-types = "0.13.3"
-reqwest = "0.12"
-secrecy = { version = "0.8.0", features = ["serde"] }
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0.114"
-sqlx = { version = "0.8.2", features = ["runtime-tokio-rustls", "postgres", "migrate", "time"] }
-strum = { version = "0.26.3", features = ["derive"] }
-thiserror = "1.0.64"
-time = "0.3"
-tokio = { version = "1.40.0", features = ["full"] }
-tonic = "0.12.2"
-tonic-reflection = "0.12.2"
-tower = "0.5.1"
-tracing = "0.1.40"
-tracing-subscriber = { version = "0.3", features = ["env-filter", "fmt", "json"] }
-
-[dev-dependencies]
-anyhow = "1.0.93"
-futures = "0.3"
-rand = "0.8"
-sha2 = "0.10"
-simple_test_case = "1.2.0"
-
-[build-dependencies]
-tonic-build = { version = "0.11", features = ["prost"] }
