--- crates/ratings_new/proto/ratings_features_chart.proto
@@ -1,57 +1,0 @@
-syntax = "proto3";
-
-package ratings.features.chart;
-
-import "ratings_features_common.proto";
-
-service Chart {
-  rpc GetChart (GetChartRequest) returns (GetChartResponse) {}
-}
-
-message GetChartRequest {
-  Timeframe timeframe = 1;
-  optional Category category = 2;
-}
-
-message GetChartResponse {
-  Timeframe timeframe = 1;
-  repeated ChartData ordered_chart_data = 2;
-  optional Category category = 3;
-}
-
-message ChartData {
-  float raw_rating = 1;
-  ratings.features.common.Rating rating = 2;
-}
-
-enum Timeframe {
-  TIMEFRAME_UNSPECIFIED = 0;
-  TIMEFRAME_WEEK = 1;
-  TIMEFRAME_MONTH = 2;
-}
-
-// The categories that can be selected, these
-// are taken directly from  `curl -sS -X GET --unix-socket /run/snapd.socket "http://localhost/v2/categories"`
-// On 2024-02-03, it may need to be kept in sync.
-enum Category {
-  ART_AND_DESIGN = 0;
-  BOOK_AND_REFERENCE = 1;
-  DEVELOPMENT = 2;
-  DEVICES_AND_IOT = 3;
-  EDUCATION = 4;
-  ENTERTAINMENT = 5;
-  FEATURED = 6;
-  FINANCE = 7;
-  GAMES = 8;
-  HEALTH_AND_FITNESS = 9;
-  MUSIC_AND_AUDIO = 10;
-  NEWS_AND_WEATHER = 11;
-  PERSONALISATION = 12;
-  PHOTO_AND_VIDEO = 13;
-  PRODUCTIVITY = 14;
-  SCIENCE = 15;
-  SECURITY = 16;
-  SERVER_AND_CLOUD = 17;
-  SOCIAL = 18;
-  UTILITIES = 19;
-}
