--- crates/ratings_new/tests/clear-db.sql
@@ -1,3 +1,0 @@
-DELETE FROM snap_categories;
-DELETE FROM users;
-DELETE FROM votes;
