+++ docker-compose.yml
@@ -23,12 +23,9 @@ services:
     ports:
       - 8080:8080
     environment:
-      RUST_LOG: "ratings_new=debug,hyper=error"
-      APP_LOG_LEVEL: "info"
-      APP_ENV: "dev"
+      RUST_LOG: "ratings=debug,hyper=error"
       APP_HOST: "0.0.0.0"
       APP_JWT_SECRET: "deadbeef"
-      APP_NAME: "ratings"
       APP_PORT: "8080"
       APP_POSTGRES_URI: "postgresql://migration_user:strongpassword@db:5432/ratings"
       APP_SNAPCRAFT_IO_URI: "http://snapcraft-mock:11111/"
@@ -36,8 +33,7 @@ services:
       APP_ADMIN_USER: "shadow"
       APP_ADMIN_PASSWORD: "maria"
     volumes:
-      # - .:/app
-      - ./crates/ratings_new:/app
+      - .:/app
       - cargo-cache:/usr/local/cargo/registry
       - target-cache:/app/target
     entrypoint: "cargo watch -i 'tests/**' -x run"
