+++ docker/prod/Dockerfile
@@ -4,7 +4,6 @@ RUN DEBIAN_FRONTEND=noninteractive apt-get update && apt-get upgrade -y && rm -r
 
 WORKDIR /app
 COPY target/release/ratings /app/ratings
-COPY sql/migrations /app/sql/migrations
 
 EXPOSE 443
 ENTRYPOINT ["/app/ratings"]
