+++ src/grpc/charts.rs
