+++ src/proto/ratings.features.chart.rs
